create and use new type for native register size
sizeof(size_t) is not guaranteed to be equal to the native register size. In
fact, it's guaranteed *not* to be equal with the x32 ABI, where the full 64-bit
register set is available, but sizeof(size_t) == 4 bytes.

Introduce and use a new type 'nint_t' that is always 64-bit wide on x86_64, and
sizeof(size_t) elsewhere.

Signed-off-by: Steven Noonan <steven@uplinklabs.net>